+++ components/index.ts
@@ -7,7 +7,6 @@ export { default as Content } from './Content';
 export { default as Slider } from './Slider';
 export { default as ImageAsset } from './ImageAsset';
 export { default as EmbedVideo } from './EmbedVideo';
-export { default as LanguageSwitcher } from './LanguageSwitcher';
 export { default as Search } from './Search';
 export { default as Categories } from './Categories';
 export { default as SocialMediaShare } from './SocialMediaShare';
